+++ src/Xteq5UserInterface/ReportCreationProgress.cs
@@ -6,22 +6,38 @@ using System.Threading.Tasks;
 
 namespace Xteq5
 {
+    public enum ReportAction
+    {
+        Starting = 1, //The report is about to be created
+        Ended = 2 //The report was created
+    }
+
     public class ReportCreationProgress
     {
         public ReportCreationProgress()
         {
-            Starting = false;
-            Ended = false;
+            this.Action = ReportAction.Starting;
         }
 
-        /// <summary>
-        /// TRUE when the report is about to be created
-        /// </summary>
-        public bool Starting { get; internal set; }
 
-        /// <summary>
-        /// TRUE when the report was created
-        /// </summary>
-        public bool Ended { get; internal set; }
+        public ReportAction Action { get; internal set; }
+
+
+        public override string ToString()
+        {
+            switch (this.Action)
+            {
+                case ReportAction.Starting:
+                    return "Creating file...";
+
+                case ReportAction.Ended:
+                    return "File created";
+
+                default:
+                    return "Unknown action";
+            }
+
+        }
+
     }
 }