+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 IMPROVEMENTS:
 
 * Added EngineVersionString to JSON File
-* Added display of currently executed script file to Xteq5GUI (using IProgress)
+* Added status of currently executed script file to Xteq5GUI
+* Added status of currently executed script file to Xteq5CLI 
 
 BUG FIXES: